Project3/requirements.txt
Faker==30.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
